test(senpi-task): split team runtime coverage by lifecycle
