+++ packages/senpi-task/src/team/runtime.test.ts
@@ -1,17 +1,12 @@
-import { existsSync } from "node:fs"
 import { stat } from "node:fs/promises"
 import { join } from "node:path"
 import { afterEach, describe, expect, test } from "bun:test"
 
-import { createRuntimeState, loadRuntimeState } from "@oh-my-opencode/team-core/team-state-store"
-
 import { readMemberTaskMap } from "./member-map"
 import { normalizeSenpiTeamSpec } from "./normalize"
-import { SenpiTeamRuntimeError, createTeam, deleteTeam } from "./runtime"
-import { toTeamCoreConfig } from "./runtime-config"
-import { resolveTeamRuntimeDirs, teamStorageBaseDir } from "./storage"
+import { createTeam } from "./runtime"
+import { resolveTeamRuntimeDirs } from "./storage"
 import {
-  FakeDestruction,
   FakeTeamManager,
   cleanupTeamRuntimeTmp,
   stateDirConfig,
@@ -240,277 +235,4 @@ describe("createTeam", () => {
     expect((await stat(worktreePath)).isDirectory()).toBe(true)
   })
 
-  test("#given a spec exceeding max_members #when created #then it is rejected before any spawn", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings({ max_members: 2 })
-    const manager = new FakeTeamManager()
-
-    // when
-    const attempt = createTeam(threeMemberSpec(), "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-    })
-
-    // then
-    await expect(attempt).rejects.toMatchObject({ code: "bounds_exceeded" })
-    expect(manager.started).toHaveLength(0)
-    expect(existsSync(join(teamStorageBaseDir(stateDir), "runtime"))).toBe(false)
-  })
-
-  test("#given the 2nd member spawn throws #when created #then the team fails and the 1st member is cancelled", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings({ max_parallel_members: 1 })
-    const manager = new FakeTeamManager({
-      behaviors: [{ kind: "ok" }, { kind: "throw", message: "spawn boom" }],
-    })
-    const spec = normalizeSenpiTeamSpec(
-      {
-        members: [
-          { name: "alpha", kind: "category", category: "quick", prompt: "a" },
-          { name: "beta", kind: "category", category: "deep", prompt: "b" },
-        ],
-      },
-      "squad",
-    )
-
-    // when
-    const attempt = createTeam(spec, "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-    })
-
-    // then
-    await expect(attempt).rejects.toBeInstanceOf(SenpiTeamRuntimeError)
-    expect(manager.cancelled.map((entry) => entry.taskId)).toEqual(["st_000001"])
-    const config = toTeamCoreConfig(settings, teamStorageBaseDir(stateDir))
-    const teamRunId = manager.started[0]?.name?.split(":")[1]
-    expect(teamRunId).toBeDefined()
-    const reloaded = await loadRuntimeState(teamRunId ?? "", config)
-    expect(reloaded.status).toBe("failed")
-  })
-
-  test("#given the member sidecar write throws #when created #then members are cancelled, the team is failed, and it never activates", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings()
-    const manager = new FakeTeamManager()
-
-    // when
-    const attempt = createTeam(threeMemberSpec(), "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-      writeMemberMap: () => Promise.reject(new Error("disk full")),
-    })
-
-    // then
-    await expect(attempt).rejects.toMatchObject({ code: "sidecar_write_failed" })
-    expect(manager.cancelled.map((entry) => entry.taskId).sort()).toEqual(["st_000001", "st_000002", "st_000003"])
-    const config = toTeamCoreConfig(settings, teamStorageBaseDir(stateDir))
-    const teamRunId = manager.started[0]?.name?.split(":")[1] ?? ""
-    const reloaded = await loadRuntimeState(teamRunId, config)
-    expect(reloaded.status).toBe("failed")
-  })
-
-  test("#given a create deadline already passed #when created #then it fails with a deadline error and no spawns", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings({ max_wall_clock_minutes: 1 })
-    const manager = new FakeTeamManager()
-    const clock = [1_000, 10_000_000]
-    let tick = 0
-    const now = () => clock[Math.min(tick++, clock.length - 1)] ?? 0
-
-    // when
-    const attempt = createTeam(threeMemberSpec(), "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-      now,
-    })
-
-    // then
-    await expect(attempt).rejects.toMatchObject({ code: "create_deadline_exceeded" })
-    expect(manager.started).toHaveLength(0)
-  })
-})
-
-describe("deleteTeam", () => {
-  async function completedResidentTeam(status: "completed" | "cancelled" = "completed") {
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings()
-    const manager = new FakeTeamManager({ defaultBehavior: { kind: "ok", status } })
-    const destruction = new FakeDestruction()
-    const spec = normalizeSenpiTeamSpec(
-      { members: [{ name: "alpha", kind: "category", category: "quick", prompt: "done" }] },
-      "squad",
-    )
-    const created = await createTeam(spec, "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-    })
-    const taskId = Object.values(created.memberTaskIds)[0]
-    if (taskId === undefined) throw new Error("expected one mapped member task")
-    const deps = { manager, destruction, stateDir, taskSettings: settings }
-    return { created, deps, manager, destruction, stateDir, taskId }
-  }
-
-  test("#given an active team #when deleted #then all member tasks are cancelled and the runtime dir is removed", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings()
-    const manager = new FakeTeamManager()
-    const spec = normalizeSenpiTeamSpec(
-      {
-        members: [
-          { name: "alpha", kind: "category", category: "quick", prompt: "a" },
-          { name: "beta", kind: "category", category: "deep", prompt: "b" },
-        ],
-      },
-      "squad",
-    )
-    const created = await createTeam(spec, "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-    })
-    const runtimeDir = resolveTeamRuntimeDirs(stateDir, created.runtimeState.teamRunId).runtimeDir
-
-    // when
-    const result = await deleteTeam(created.runtimeState.teamRunId, {
-      manager,
-      destruction: new FakeDestruction(),
-      stateDir,
-      taskSettings: settings,
-    })
-
-    // then
-    expect([...result.cancelledTaskIds].sort()).toEqual(["st_000001", "st_000002"])
-    expect(manager.cancelled).toHaveLength(2)
-    expect(existsSync(runtimeDir)).toBe(false)
-  })
-
-  test("#given a team still in creating #when deleted #then an invalid-state error is thrown", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings()
-    const manager = new FakeTeamManager()
-    const spec = normalizeSenpiTeamSpec(
-      { members: [{ name: "alpha", kind: "category", category: "quick", prompt: "a" }] },
-      "squad",
-    )
-    const config = toTeamCoreConfig(settings, teamStorageBaseDir(stateDir))
-    const seeded = await createRuntimeState(spec, "lead-session", "project", config)
-
-    // when
-    const attempt = deleteTeam(seeded.teamRunId, {
-      manager,
-      destruction: new FakeDestruction(),
-      stateDir,
-      taskSettings: settings,
-    })
-
-    // then
-    await expect(attempt).rejects.toMatchObject({ code: "invalid_delete_state" })
-  })
-
-  test("#given an already-deleted team #when deleted again #then it is a no-op", async () => {
-    // given
-    const stateDir = stateDirConfig(tempProjectDir())
-    const settings = taskSettings()
-    const manager = new FakeTeamManager()
-    const spec = normalizeSenpiTeamSpec(
-      { members: [{ name: "alpha", kind: "category", category: "quick", prompt: "a" }] },
-      "squad",
-    )
-    const created = await createTeam(spec, "project", {
-      manager,
-      stateDir,
-      taskSettings: settings,
-      leadSessionId: "lead-session",
-      spawnDepth: 1,
-    })
-    const deps = { manager, destruction: new FakeDestruction(), stateDir, taskSettings: settings }
-    await deleteTeam(created.runtimeState.teamRunId, deps)
-
-    // when
-    const second = await deleteTeam(created.runtimeState.teamRunId, deps)
-
-    // then
-    expect(second.cancelledTaskIds).toEqual([])
-  })
-
-  test("#given a completed resident member #when deleted #then cancellation stays noop while lifecycle destroys it and the runtime is removed", async () => {
-    // given
-    const { created, deps, manager, destruction, stateDir, taskId } = await completedResidentTeam()
-    const runtimeDir = resolveTeamRuntimeDirs(stateDir, created.runtimeState.teamRunId).runtimeDir
-
-    // when
-    const result = await deleteTeam(created.runtimeState.teamRunId, deps)
-
-    // then
-    expect(manager.cancelled).toEqual([{ taskId, reason: `delete team ${created.runtimeState.teamRunId}` }])
-    expect(result.cancelledTaskIds).toEqual([])
-    expect(destruction.calls).toEqual([{ taskId, cause: "cancel" }])
-    expect(manager.get(taskId)?.status).toBe("completed")
-    expect(existsSync(runtimeDir)).toBe(false)
-  })
-
-  test("#given a completed resident member already disposed #when deleted #then destruction is not repeated", async () => {
-    // given
-    const { created, deps, manager, destruction, taskId } = await completedResidentTeam()
-    manager.setResidency(taskId, "disposed")
-
-    // when
-    const result = await deleteTeam(created.runtimeState.teamRunId, deps)
-
-    // then
-    expect(result.cancelledTaskIds).toEqual([])
-    expect(destruction.calls).toEqual([])
-  })
-
-  test("#given a resident member already cancelled by an in-flight cancellation #when deleted #then destruction is not raced", async () => {
-    // given
-    const { created, deps, manager, destruction } = await completedResidentTeam("cancelled")
-
-    // when
-    const result = await deleteTeam(created.runtimeState.teamRunId, deps)
-
-    // then
-    expect(result.cancelledTaskIds).toEqual([])
-    expect(destruction.calls).toEqual([])
-  })
-
-  test("#given two delete calls for one team #when invoked concurrently #then they share one cancellation and one destruction", async () => {
-    // given
-    const { created, deps, manager, destruction } = await completedResidentTeam()
-
-    // when
-    const [first, second] = await Promise.all([
-      deleteTeam(created.runtimeState.teamRunId, deps),
-      deleteTeam(created.runtimeState.teamRunId, deps),
-    ])
-
-    // then
-    expect(first).toEqual(second)
-    expect(manager.cancelled).toHaveLength(1)
-    expect(destruction.calls).toHaveLength(1)
-  })
 })
